+++ src/main/java/com/susanghan_guys/server/user/application/UserAuthService.java
@@ -6,9 +6,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.susanghan_guys.server.global.common.code.ErrorCode;
 import com.susanghan_guys.server.global.exception.BusinessException;
 import com.susanghan_guys.server.global.jwt.JwtProvider;
-import com.susanghan_guys.server.global.oauth2.application.TokenBlackListService;
-import com.susanghan_guys.server.global.oauth2.domain.RefreshToken;
-import com.susanghan_guys.server.global.oauth2.infrastructure.persistence.RefreshTokenRepository;
+import com.susanghan_guys.server.oauth2.application.TokenBlackListService;
+import com.susanghan_guys.server.oauth2.domain.RefreshToken;
+import com.susanghan_guys.server.oauth2.infrastructure.persistence.RefreshTokenRepository;
 import com.susanghan_guys.server.global.util.RedisUtil;
 import com.susanghan_guys.server.user.domain.User;
 import com.susanghan_guys.server.user.dto.response.RefreshTokenResponse;
